reorganizing; warn about orphaned keysyms
I moved the HTML boilerplate into its own file.
I moved the config stuff to the head of the script.
I added a check to make sure that the keysyms mentioned in the i3config
also appear in the specified layout, giving a warning in stderr and the
HTML output if they don't.